pyproject.toml
[project]
name = "testaiownik"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "docx>=0.2.4",
  "dotenv>=0.9.9",
  "langchain>=0.3.25",
  "langchain-openai>=0.3.21",
  "langgraph>=0.4.8",
  "logging>=0.4.9.6",
  "numpy>=2.3.0",
  "opencensus-ext-azure>=1.1.15",
  "pathlib>=1.0.1",
  "pdfplumber>=0.11.7",
  "pydantic>=2.11.5",
  "python-docx>=1.2.0",
  "python-pptx>=1.0.2",
  "qdrant-client>=1.14.3",
  "requests>=2.32.3",
  "trustcall>=0.0.39",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "colorlog>=6.9.0",
    "coverage>=7.9.0",
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.13",
]


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = "test_*.py"
addopts = "--cov=src/Testaiownik --cov-report=term-missing -v"

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
packages = ["Testaiownik"]
explicit_package_bases = true
